+++ lib/cluster/pkgIndex.tcl
@@ -9,4 +9,5 @@
 # full path name of this file's directory.
 
 package ifneeded cluster 0.2 [list source [file join $dir cluster.tcl]]
+package ifneeded cluster::swarm 0.2 [list source [file join $dir swarm.tcl]]
 package ifneeded cluster::virtualbox 0.1 [list source [file join $dir virtualbox.tcl]]